moving AddItemForm to new file and fixing double message bug
